- Termino de tutorial sobre control de conflictos de simultaneidad

+++ EFCoreMVC/Controllers/DepartmentsController.cs
@@ -105,11 +105,12 @@ namespace EFCoreMVC.Controllers
 
             var departmentToUpdate = await _context.Departments.Include(i => i.Administrator).FirstOrDefaultAsync(m => m.DepartmentID == id);
 
+            // Otro usuario eliminó el departamento
             if (departmentToUpdate == null)
             {
                 Department deletedDepartment = new Department();
 
-                await TryToUpdateAsync(deletedDepartment);
+                await TryUpdateModelAsync(deletedDepartment);
 
                 ModelState.AddModelError(string.Empty,
                     "Unable to save changes. The department was deleted by another user.");
@@ -131,6 +132,8 @@ namespace EFCoreMVC.Controllers
                 }
                 catch (DbUpdateConcurrencyException ex)
                 {
+
+                    // El código del bloque catch de esa excepción obtiene la entidad Department afectada que tiene los valores actualizados de la propiedad Entries del objeto de excepción.
                     var exceptionEntry = ex.Entries.Single();
                     var clientValues = (Department)exceptionEntry.Entity;
                     var databaseEntry = exceptionEntry.GetDatabaseValues();
@@ -170,6 +173,7 @@ namespace EFCoreMVC.Controllers
                                 "have been displayed. If you still want to edit this record, click " +
                                 "the Save button again. Otherwise click the Back to List hyperlink");
 
+                        // Por último, el código establece el valor RowVersion de departmentToUpdate para el nuevo valor recuperado de la base de datos. Este nuevo valor RowVersion se almacenará en el campo oculto cuando se vuelva a mostrar la página Edit y, la próxima vez que el usuario haga clic en Save, solo se detectarán los errores de simultaneidad que se produzcan desde que se vuelva a mostrar la página Edit.
                         departmentToUpdate.RowVersion = (byte[])databaseValues.RowVersion;
 
                         ModelState.Remove("RowVersion");
@@ -183,7 +187,7 @@ namespace EFCoreMVC.Controllers
         }
 
         // GET: Departments/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? concurrencyError)
         {
             if (id == null)
             {
@@ -191,25 +195,57 @@ namespace EFCoreMVC.Controllers
             }
 
             var department = await _context.Departments
-                .Include(d => d.Administrator)
-                .FirstOrDefaultAsync(m => m.DepartmentID == id);
+                                                .Include(d => d.Administrator)
+                                                .AsNoTracking()
+                                                .FirstOrDefaultAsync(m => m.DepartmentID == id);
             if (department == null)
             {
+                if (concurrencyError.GetValueOrDefault())
+                {
+                    return RedirectToAction(nameof(Index));
+                }
+
                 return NotFound();
+            }
+
+            if (concurrencyError.GetValueOrDefault())
+            {
+                ViewData["ConcurrencyErrorMessage"] = "The record you attempted to delete " +
+                    "was modifiedd by another user after you got the original values. " +
+                    "The delete operation was canceled and the current values in the " +
+                    "database have been displayed. If you still want to delete this " +
+                    "record, click the Delete button again. Otherwise " +
+                    "click the Back to List hyperlink.";
             }
 
             return View(department);
         }
 
+
         // POST: Departments/Delete/5
-        [HttpPost, ActionName("Delete")]
+        [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> Delete(Department department)
+        // Ha cambiado este parámetro por una instancia de la entidad Department creada por el enlazador de modelos. Esto proporciona a EF acceso al valor de la propiedad RowVersion, además de la clave de registro.
         {
-            var department = await _context.Departments.FindAsync(id);
-            _context.Departments.Remove(department);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            try
+            {
+                // Si ya se ha eliminado el departamento, el método AnyAsync devuelve false y la aplicación simplemente vuelve al método de índice.
+                if (await _context.Departments.AnyAsync(m => m.DepartmentID == department.DepartmentID))
+                {
+                    _context.Departments.Remove(department);
+                    await _context.SaveChangesAsync();
+                }
+
+                return RedirectToAction(nameof(Index));
+            }
+            // Si se detecta un error de simultaneidad, el código vuelve a mostrar la página de confirmación de Delete y proporciona una marca que indica que se debería mostrar un mensaje de error de simultaneidad.
+            catch (DbUpdateConcurrencyException ex)
+            {
+                // Log the error
+                return RedirectToAction(nameof(Delete), new { concurrencyError = true, id = department.DepartmentID });
+            }
+
         }
 
         private bool DepartmentExists(int id)
